Adjust alignment and spacing of label, input, and child fields of `FieldContainer` and `Field` (#2370)
Previously, the label and input of fields were centered vertically. 
Now, the label is positioned from the top, and each input has a minimum
height of 40px.

Additionally, the spacing of child fields was adjusted to make sure
elements with less height, such as radios and checkboxes, would still be
centered vertically with the label and would still be aligned with their
siblings and inputs of other fields.