Add eslint rule to restrict barrel react imports (#2484)
Use named imports instead of barrel importing React. Doing so will
result in less code and better readability. The new [React
docs](https://react.dev/) also use named imports and never import React
itself.

To avoid conflicts, we temporarily add the eslint rules in the packages
on main where it is possible. The restrict-import rule will be added in
the v8 (next) branch in the eslint-config package, then they will be
removed there again.

We can not add it on main for the `demo/admin` and the `admin/cms-admin`
package because of the generator. The generated files can not be
modified without a breaking change and will be done in next/v8.