`BuildsService`: Start all jobs that match the scope exactly (#2486)
Previously, the first job that matched the scope exactly would be
started, and the rest would be ignored. This has been fixed so that all
jobs that match the scope exactly are started.

This bug occurred in one of our client's projects where we have two
frontends: the site and a newsletter frontend. Both use the same scope:

<img width="479" alt="Bildschirmfoto 2024-08-29 um 14 16 11"
src="https://github.com/user-attachments/assets/2c27fd64-6897-4e20-bd77-6df5aaad89a8">

If a change was detected, only the newsletter frontend was built but the
site wasn't. This lead to cases where it took 24h until changes were
visible on the site (because everything is rebuilt every 24h).