Admin Generator (Future): Radio select structure rework (#2430)
This PR proposes changes to the decision logic for static select
generation to better reflect the design guidelines.

Note: This PR is based on #2378.

**Quoted design guidelines for static selects:**
_When to use a radio button:_
- _Only use if it is mandatory AND has a maximum of 5 options to choose_

_When to use a select field:_
- _Always, if it is an optional value_
- _Always, if there are more than 5 options_

**Code Before:**
The logic is solely decided by setting the `inputType`. The developer
has to know exactly when to use which variant of static select.

```
// inputType = "select" | "radio" | undefined -> set in FormFieldConfig

if (config.inputType === "radio") {
    // radio
} else {
    // select
}
```

**Code After:**
This logic takes current design guidelines into account. The developer
does not have to remember the design details.
Setting `inputType` allows to manually override the guidelines.

```
// inputType = "select" | "radio" | undefined -> set in FormFieldConfig

const renderAsRadio = config.inputType === "radio" || (required && values.length <= 5 && config.inputType !== "select");
if (renderAsRadio) {
    // radio
} else if (!required || values.length > 5 || config.inputType === "select") {
    // select
}
```