Add support for a native Badge in `MenuItemRouterLink` and correctly format `MenuItemAnchorLink` (#2431)
This adds a `badgeContent` prop to the `MenuItemRouterLink` component,
which automatically renders a correctly formatted badge. It can be
passed via prop directly to a component or by passing it as property to
the object in `masterMenuData`. Check out the changeset for example
usages.

Additionally the `MenuItemAnchorLink` has been extended to automatically
render and correctly format a `LinkExternal` icon.