Create Site Preview JWT in the API (#2554)
Allows having the Site Preview unauthenticated

## Description

Before: the JWT with the preview scope and setting was generated in the
site. To check if the current user is allowed to preview the given scope
the site had to make a request to the API and therefore needed the
access token of the current user. With this setup the site has to be
behind an authproxy.

Now: the JWT is generated by the API and submitted to the site. This way
the site can be public.

## BREAKING
The API now requires the SITE_PREVIEW_SECRET environment variable. To
make it explicit to add it this env now is also mandatory for local
development.