Admin Generator (Future): Add  `FinalFormRangeInput` to form generator (#2612)
`FinalFormRangeInput` fields can now be generated in forms. The new
field type is `numberRange`. Data of the field is saved in this format:
`{ min: number, max: number }`

The names `min` and `max` are currently hardcoded, as they **have to**
be named this way. `FinalFormRangeInput` passes the values named like
this.