COM-957: Add basic warning entity (#2503)
In the Warning Module there should be a warning entity where warnings
are saved.
Entries into this table should later be added with a validator or via a
block, maybe even manually.
This is the start of the central warning system in order to have a basic
table and to be able to start with the validator afterwards.