Fix Demo Site Pages (#2695)
The legacyPagesRouterSitePreviewApiHandler used parts from SitePreviewUtils which had a server-only restriction. However, the server-only package can only be used in RSC compatible setups (i.e. the App Router). To fix this, we move the App Router specific parts to a separate file, ensuring that SitePreviewUtils can be used for both Pages Router and App Router.