Prevent providing a `User` type in the GraphQL schema (#2755)
We currently export a `User` type in the GraphQL schema. If a web defines this type by itself this leads to a naming collision. Comet should not do such dominant naming, so this PR renames some type to mitigate this issue:
- `User` => `UserPermissionsUser `
- `UserFilter`/`UserSort` => Prefix with `UserPermissions`

The renaming only affects the GraphQL schema! The `User` typescript type, which is also used in projects, remains as is. However, a code based naming conflict can be solved much more easily.