Fix showing only the first story in docs
Previously each canvas would show the same story on a given page, the
first one that was loaded. This is fixed by only rendering the
`Canvas` directly, instead of a `Story`, nested within a `Canvas`.

This seems to be the correct way to render stories, according to the
Storybook docs: https://storybook.js.org/docs/writing-docs/mdx#basic-example