Prevent `ContentScopeIndicator` from crashing when a scope part is `undefined` (#2859)
Normally, no part of the scope should be `undefined`. It should either have a value or be non-existent. Still, I think in practice it will happen that scope parts become undefined for some reason. 

So this PR makes the ContentScopeIndicator more resistant. Instead of crashing if a scope part is `undefined`, it now just ignores it.