Implement adding, deleting, editing and dragging of `TableBlock` rows and columns (#2808)
The position of rows and columns is no longer stored as a separate field in the object, the objects position in the array is used instead. 

Using `onColumnOrderChange` in the same way as `onRowOrderChange` would seem logical, but caused issues when updating the state while the columns were being dragged. That's why we subscribe to DataGrids `columnHeaderDragEnd` event instead to update the state with the new column position.