Prevent unintended layout shift after initial render of `FieldContainer`  (#2872)
`FieldContainer` and, therefore, all `Field` components can be set to use the `horizontal` variant to display the label and input next to each other instead of below each other, as in the default `vertical` variant. 

To make fields work better on mobile, we don't apply the styling of the `horizontal` variant until the field has at least `600px` of space available, see https://github.com/vivid-planet/comet/pull/1755.

Previously, this was done using the `useObservedWidth` hook to get the width of the `Root` slot using a ref. 
This caused some issues, as the refs `current` element would sometimes be `null` on the component's initial render and would take the duration of one `ResizeObserver` debounce to be set correctly, causing unintended layout shift, e.g., when switching to a new tab that contains a `FieldContainer`. 

By using CSS container queries, the correct styling can be applied as soon as the element is rendered. 

Co-authored-by: Johannes Obermair <48853629+johnnyomair@users.noreply.github.com>