Don't render a tooltip around disabled buttons
This doesn't change the behavior, as tooltips can't listen to events on disabled elements anyway.
This prevents the following error in the console:

```
MUI: You are providing a disabled `button` child to the Tooltip component.
A disabled element does not fire events.
Tooltip needs to listen to the child element's events to display the title.
```