Admin Generator: Fix Form Textfield Update to an empty string
Work around final-form 'feature' that uses undefined for empty text inputs (https://final-form.org/docs/final-form/faq#why-does-final-form-set-my--field-value-to-undefined) causing update mutation
to leave the field unchanged (as undefined is sent - which means basically the field is not sent)

Solution is to add a formValues.field ?? '', which changes undefined to ''.

In Demo, to reproduce this issue, change the description field to non-required, allowing an empty string. (still not nullable in database though)