Fix keyboard input of MUI DatePicker
Previously, applying the value on every keystroke would cause
unexpected behaivor while typing, e.g., changing month to "10" would
show in "01" in the input, while still applying "10".

This is fixed by only updating an internal value during keyboard
input (`onChange`), then applying it on blur of the input, after
typing is complete.

When selecting a value in the picker, the value is still applied
immediately, with `onAccept`.