Mark Button `color` prop as deprecated
The `ButtonProps` type already does not include the `color` prop.
Due to the usage of the type-cast `as OverridableComponent` on the Button component, the `color` prop is added again, from the html elements props.

A changeset is likely not necessary, as the removal of the `color` prop and the necessary changes have already been mentioned in a previous changeset.

A cleaner solution would be to remove the type-cast all together and fix the underlying issue.
This will be attempted in a separate PR into `next` to avoid a breaking change.