Always render the cancel button on the left
If `DialogActions` only has one item, that item is aligned to the right.
Therefore the cancel button's alignment would be inconsistent, depending on the ability to edit the form.

An alternative would be to always render the save button and use the `disabled` prop but the current fix has less visual impact.